Map given preset label with labels from server to get ID
With this change it is possible to define a label in the checkmarx options instead of only an id.
The given label will be looked up from the available labels on the checkmarx server.
If the label is found the respective id will be used. Otherwise the default value 7 will be used.