Update hudson.remoting.ClassFilter
Added Results class to the class filter as during deserialization this was failing. And it is serialized so as to work in master slave ode if during scan any exception occurs , the exception was not getting added to results object. Hence failTheBuild() in CxScanBuilder class not not making the build fail/unstable . This issue was discovered during defect AB#2334. The issue was even if user gives wrong user name and password, the job still continue to be SUCCESS in master slave mode